+++ components/browser/awesomebar/src/main/java/mozilla/components/browser/awesomebar/BrowserAwesomeBar.kt
@@ -137,7 +137,7 @@ class BrowserAwesomeBar @JvmOverloads constructor(
         job?.cancel()
 
         job = scope.launch {
-            suggestionsAdapter.optionallyClearSuggestions()
+            suggestionsAdapter.removeAllSuggestions()
 
             providers.forEach { provider ->
                 launch {