+++ components/browser/awesomebar/src/test/java/mozilla/components/browser/awesomebar/SuggestionsAdapterTest.kt
@@ -33,12 +33,14 @@ class SuggestionsAdapterTest {
     fun `addSuggestions() should add suggestions of provider`() {
         val adapter = SuggestionsAdapter(mock())
 
+        val provider: AwesomeBar.SuggestionProvider = mock()
+
         val suggestions = listOf<AwesomeBar.Suggestion>(
             mock(), mock(), mock())
 
         assertEquals(0, adapter.itemCount)
 
-        adapter.addSuggestions(mockProvider(), suggestions)
+        adapter.addSuggestions(provider, suggestions)
 
         assertEquals(3, adapter.itemCount)
         assertEquals(suggestions, adapter.suggestions)
@@ -48,8 +50,7 @@ class SuggestionsAdapterTest {
     fun `addSuggestions() always clears previous suggestions of provider`() {
         val adapter = SuggestionsAdapter(mock())
 
-        val provider = mockProvider()
-        `when`(provider.shouldClearSuggestions).thenReturn(true)
+        val provider: AwesomeBar.SuggestionProvider = mock()
 
         assertEquals(0, adapter.itemCount)
 
@@ -61,21 +62,12 @@ class SuggestionsAdapterTest {
         adapter.addSuggestions(provider, suggestions)
         assertEquals(3, adapter.itemCount)
         assertEquals(suggestions, adapter.suggestions)
-
-        // shouldClearSuggestions is used to indicate whether or not
-        // suggestions should be cleared right away on input changes.
-        // When we're adding newly computed suggestions, we should
-        // always clear old ones to prevent duplicates.
-        `when`(provider.shouldClearSuggestions).thenReturn(false)
-        adapter.addSuggestions(provider, suggestions)
-        assertEquals(3, adapter.itemCount)
-        assertEquals(suggestions, adapter.suggestions)
     }
 
     @Test
     fun `removeSuggestions() should remove suggestions of provider`() {
         val adapter = SuggestionsAdapter(mock())
-        val provider = mockProvider()
+        val provider: AwesomeBar.SuggestionProvider = mock()
         val suggestions = listOf<AwesomeBar.Suggestion>(
                 mock(), mock(), mock())
 
@@ -91,45 +83,25 @@ class SuggestionsAdapterTest {
     fun `clearSuggestions removes suggestions from adapter`() {
         val adapter = SuggestionsAdapter(mock())
 
-        adapter.addSuggestions(mockProvider(), listOf(
+        val provider: AwesomeBar.SuggestionProvider = mock()
+
+        adapter.addSuggestions(provider, listOf(
             mock(), mock(), mock()))
 
         assertEquals(3, adapter.itemCount)
 
-        adapter.optionallyClearSuggestions()
+        adapter.removeAllSuggestions()
 
         assertEquals(0, adapter.itemCount)
-    }
-
-    @Test
-    fun `clearSuggestions does not remove suggestions if provider has set shouldClearSuggestions to false`() {
-        val adapter = SuggestionsAdapter(mock())
-
-        adapter.addSuggestions(mockProvider(shouldClearSuggestions = false), listOf(
-            mock(), mock(), mock()))
-
-        assertEquals(3, adapter.itemCount)
-
-        adapter.optionallyClearSuggestions()
-
-        assertEquals(3, adapter.itemCount)
-
-        adapter.addSuggestions(mockProvider(shouldClearSuggestions = true), listOf(
-            mock(), mock(), mock(), mock()
-        ))
-
-        assertEquals(7, adapter.itemCount)
-
-        adapter.optionallyClearSuggestions()
-
-        assertEquals(3, adapter.itemCount)
     }
 
     @Test
     fun `Suggestions are getting ordered by weight descending`() {
         val adapter = SuggestionsAdapter(mock())
 
-        adapter.addSuggestions(mockProvider(), listOf(
+        val provider: AwesomeBar.SuggestionProvider = mock()
+
+        adapter.addSuggestions(provider, listOf(
             AwesomeBar.Suggestion(mock(), title = "Hello", score = 10),
             AwesomeBar.Suggestion(mock(), title = "World", score = 2),
             AwesomeBar.Suggestion(mock(), title = "How", score = 7),
@@ -156,7 +128,9 @@ class SuggestionsAdapterTest {
     fun `Adapter uses different view holder for suggestions with chips`() {
         val adapter = SuggestionsAdapter(mock())
 
-        adapter.addSuggestions(mockProvider(), listOf(
+        val provider: AwesomeBar.SuggestionProvider = mock()
+
+        adapter.addSuggestions(provider, listOf(
             AwesomeBar.Suggestion(mock(), title = "Test"),
             AwesomeBar.Suggestion(mock(), title = "World", chips = listOf(
                 AwesomeBar.Suggestion.Chip("Chip1"),
@@ -199,11 +173,13 @@ class SuggestionsAdapterTest {
 
         val suggestion: AwesomeBar.Suggestion = mock()
 
+        val provider: AwesomeBar.SuggestionProvider = mock()
+
         val viewHolder: SuggestionViewHolder = mock()
         val wrapper = ViewHolderWrapper(viewHolder, mock())
         doReturn(wrapper).`when`(adapter).onCreateViewHolder(any(), anyInt())
 
-        adapter.addSuggestions(mockProvider(), listOf(suggestion))
+        adapter.addSuggestions(provider, listOf(suggestion))
 
         adapter.onBindViewHolder(wrapper, 0)
 
@@ -259,13 +235,5 @@ class SuggestionsAdapterTest {
         adapter.onViewRecycled(wrapper)
 
         verify(viewHolder).recycle()
-    }
-
-    private fun mockProvider(
-        shouldClearSuggestions: Boolean = true
-    ): AwesomeBar.SuggestionProvider {
-        val provider: AwesomeBar.SuggestionProvider = mock()
-        doReturn(shouldClearSuggestions).`when`(provider).shouldClearSuggestions
-        return provider
     }
 }