+++ components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/ClipboardSuggestionProvider.kt
@@ -55,10 +55,6 @@ class ClipboardSuggestionProvider(
             }
         ))
     }
-
-    override val shouldClearSuggestions: Boolean
-        // We do not want the suggestion of this provider to disappear and re-appear when text changes.
-        get() = false
 }
 
 private fun findUrl(text: String): String? {