+++ components/feature/awesomebar/src/test/java/mozilla/components/feature/awesomebar/provider/HistoryStorageSuggestionProviderTest.kt
@@ -12,7 +12,6 @@ import mozilla.components.concept.storage.VisitType
 import mozilla.components.support.test.eq
 import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.mockito.Mockito.`when`
@@ -50,12 +49,6 @@ class HistoryStorageSuggestionProviderTest {
 
         val suggestions = provider.onInputChanged("moz")
         assertEquals(20, suggestions.size)
-    }
-
-    @Test
-    fun `Provider suggestion should not get cleared when text changes`() {
-        val provider = HistoryStorageSuggestionProvider(mock(), mock())
-        assertFalse(provider.shouldClearSuggestions)
     }
 
     @Test